Manage exception per Platform
