+++ lib/src/browser/dart_wrappers.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 
 import 'package:js/js_util.dart';
-import 'package:webthree/src/core/exception.dart';
+import 'package:webthree/src/core/exception_utils_js.dart'
+    if (dart.library.io) 'package:webthree/src/core/exception_utils_js.dart'
+    if (dart.library.js) 'package:webthree/src/core/exception_utils_js.dart';
 
 import '../../credentials.dart';
 import '../../json_rpc.dart';
